refactor: split document mutation policy helpers
Refs #657
